Fix typo (#1819)
* Fix typo

I'm pretty sure you meant to write here `md-alignment-[vertical]-[horizontal]`, but in the docs it says `md-alignment-[horizontal]-[vertical]`. That threw me off for a bit.

Overall, this is a super awesome package! I love working with it!!

* fix typo in api section

I noticed that at the bottom of the page, in the API section, the order also needs to be reversed. "x" should refer to horizontal, and "y" should refer to vertical.

Co-authored-by: dragosct <17456857+dragosct10@users.noreply.github.com>